requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
scikit-learn==1.5.2
joblib==1.4.2
pyyaml==6.0.2
httpx==0.27.2
pytest==8.3.3
numpy==1.26.4
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7
python-dotenv==1.0.1
ruff==0.6.9
